refactor: refactor all the controller to service
