+++ services/tweet-services.js
@@ -1,7 +1,9 @@
-const { Tweet, User, Like, Reply } = require('../models')
+const helpers = require('../_helpers')
+
 const { relativeTimeFromNow } = require('../helpers/dayjs-helpers')
 const { getUserData } = require('../helpers/getUserData')
-const helpers = require('../_helpers')
+const { Tweet, User, Like, Reply } = require('../models')
+
 const tweetServices = {
   getTweets: async (req, cb) => {
     try {
@@ -65,7 +67,7 @@ const tweetServices = {
           createdAt: relativeTimeFromNow(tweet.createdAt)
         })
       })
-      .catch(err => cb(err))
+      .catch(err => cb (err))
   },
   postTweets: async (req, cb) => {
     try {
@@ -86,13 +88,54 @@ const tweetServices = {
       })
       const tweetData = tweet.toJSON()
         tweetData.createdAt = relativeTimeFromNow(tweetData)
-      return cb(null, {
+      return cb (null, {
         status: 'success',
         message: '登入成功！',
         tweetData
       })
     } catch (err) {
-      cb(err)
+      cb (err)
+    }
+  },
+  addLike: async(req, cb) => {
+    try{
+      const { id } = req.params
+      const tweet = await Tweet.findByPk(id)
+      const like = await Like.findOne({
+              where: {
+                UserId: helpers.getUser(req).id,
+                TweetId: id
+              }
+            })
+      if (tweet.id !== Number(id)) throw new Error("Tweet不存在!")
+      if (like) throw new Error('你已經like過這篇Tweet了')
+
+      const likeCreate = await Like.create({
+        UserId: helpers.getUser(req).id,
+        TweetId: Number(id)
+      })
+      cb (null, likeCreate)
+    } catch (err) {
+          cb (err)
+    }
+  },
+  removeLike: async (req, cb) => {
+    try {
+      const { id } = req.params
+      const like = await Like.findOne({
+        where: {
+          UserId: helpers.getUser(req).id,
+          tweetId: id
+        }
+      })
+      if (!like) {
+        throw new Error('這篇Tweet沒被like')
+      }
+
+      await like.destroy()
+      cb (null,{ message: 'Like 取消成功' })
+    } catch (err) {
+      cb (err)
     }
   }
 }
